Refactor: Factory: Move the ansible-related code to a separate factory class
